Fix/#38 fix fuking shjit

+++ app/src/main/java/com/hsLink/hslink/presentation/mypage/navigation/career/CareerEditNavigation.kt
@@ -10,7 +10,7 @@ import com.hsLink.hslink.presentation.mypage.screen.career.CareerEditRoute
 import kotlinx.serialization.Serializable
 
 // ← careerId 파라미터 추가
-fun NavController.navigateToCareerEdit(careerId: Long, navOptions: NavOptions? = null) {
+fun NavController.navigateToCareerEdit(careerId: Int, navOptions: NavOptions? = null) {
     navigate(CareerEdit(careerId = careerId), navOptions)
 }
 
@@ -20,18 +20,16 @@ fun NavGraphBuilder.careerNavGraph(
 ) {
     composable<CareerEdit> { backStackEntry ->
         val careerEdit = backStackEntry.arguments?.let {
-            // ← careerId 추출
-            CareerEdit(careerId = it.getLong("careerId"))
-        } ?: CareerEdit(careerId = 0L)
+            CareerEdit(careerId = it.getInt("careerId")) // ← getLong → getInt 변경!
+        } ?: CareerEdit(careerId = 0) // ← 0L → 0 변경!
 
         CareerEditRoute(
             paddingValues = padding,
             navController = navController,
-            careerId = careerEdit.careerId // ← careerId 전달
+            careerId = careerEdit.careerId.toLong() // ← CareerEditRoute가 Long을 받으면 여기서 변환
         )
     }
 }
 
-// ← data object에서 data class로 변경
 @Serializable
-data class CareerEdit(val careerId: Long) : Route
+data class CareerEdit(val careerId: Int) : Route
